mywebsite/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-ckeditor==6.0.0
django-crispy-forms==1.11.0
django-embed-video==1.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.3.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.59
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
